requirements.txt
aiohttp~=3.8.3
motor~=3.1.1
python-telegram-bot[job-queue, webhooks]==20.0
pytz~=2022.7.1